Update service tests

+++ src/main/kotlin/com/simkord/fridgepalbackend/service/ProductService.kt
@@ -3,11 +3,13 @@ package com.simkord.fridgepalbackend.service
 import com.github.michaelbull.result.*
 import com.simkord.fridgepalbackend.application.request.ProductFilters
 import com.simkord.fridgepalbackend.datasource.ProductDataSource
+import com.simkord.fridgepalbackend.datasource.database.entity.ProductEntity
 import com.simkord.fridgepalbackend.service.mapper.*
 import com.simkord.fridgepalbackend.service.model.Product
 import com.simkord.fridgepalbackend.service.model.ServiceError
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.multipart.MultipartFile
 
 @Service
@@ -32,25 +34,33 @@ class ProductService(
         return productDataSource.saveProduct(product.toProductEntity()).mapToServiceResult { it.toProduct() }
     }
 
+    @Transactional
     fun deleteProduct(productId: Long): Result<Unit, ServiceError> {
-        checkProductExistsById(productId).onFailure {
-            return@deleteProduct Err(it)
+        val existingProduct = getExistingProduct(productId).fold(
+            success = { it },
+            failure = { return@deleteProduct Err(it) },
+        )
+
+        productDataSource.deleteProductById(productId).fold(
+            success = {},
+            failure = { return@deleteProduct Err(it.toServiceError()) },
+        )
+
+        existingProduct.imageId?.let { imageId ->
+            cloudinaryService.deleteImageFromCloudinary(imageId).fold(
+                success = {},
+                failure = { return@deleteProduct Err(it) },
+            )
         }
 
-        return productDataSource.deleteProductById(productId).mapToServiceResult { }
+        return Ok(Unit)
     }
 
     fun updateProductImage(image: MultipartFile, productId: Long): Result<Product, ServiceError> {
-        val product = productDataSource.getProductById(productId).fold(
+        val existingProduct = getExistingProduct(productId).fold(
             success = { it },
-            failure = { return@updateProductImage Err(it.toServiceError()) },
+            failure = { return@updateProductImage Err(it) },
         )
-
-        verifyProductExists(product.isPresent, productId).onFailure {
-            return@updateProductImage Err(it)
-        }
-
-        val existingProduct = product.get()
 
         val uploadedImage = cloudinaryService.uploadImageToCloudinary(image).fold(
             success = { it },
@@ -60,7 +70,13 @@ class ProductService(
         existingProduct.imageUrl = uploadedImage.imageUrl
         existingProduct.imageId = uploadedImage.imageId
 
-        return productDataSource.saveProduct(existingProduct).mapToServiceResult { it.toProduct() }
+        val result = productDataSource.saveProduct(existingProduct).mapToServiceResult { it.toProduct() }
+
+        if (result.isErr && uploadedImage.imageId != null) {
+            cloudinaryService.deleteImageFromCloudinary(uploadedImage.imageId)
+        }
+
+        return result
     }
 
     private fun checkProductExistsById(productId: Long): Result<Unit, ServiceError> {
@@ -71,7 +87,27 @@ class ProductService(
         return verifyProductExists(productExists, productId)
     }
 
-    fun deleteProductImage(imageId: String): Result<Unit, ServiceError> {
+    @Transactional
+    fun deleteProductImage(productId: Long): Result<Unit, ServiceError> {
+        val existingProduct = getExistingProduct(productId).fold(
+            success = { it },
+            failure = { return@deleteProductImage Err(it) },
+        )
+
+        if (existingProduct.imageId == null) {
+            return Err(ServiceError(HttpStatus.NOT_FOUND.value(), "Existing image not found for product $productId"))
+        }
+
+        val imageId = existingProduct.imageId!!
+
+        existingProduct.imageId = null
+        existingProduct.imageUrl = null
+
+        productDataSource.saveProduct(existingProduct).fold(
+            success = { it },
+            failure = { return@deleteProductImage Err(it.toServiceError()) },
+        )
+
         return cloudinaryService.deleteImageFromCloudinary(imageId).fold(
             success = { Ok(Unit) },
             failure = { Err(it) },
@@ -84,5 +120,18 @@ class ProductService(
         }
 
         return Ok(Unit)
+    }
+
+    private fun getExistingProduct(productId: Long): Result<ProductEntity, ServiceError> {
+        val product = productDataSource.getProductById(productId).fold(
+            success = { it },
+            failure = { return@getExistingProduct Err(it.toServiceError()) },
+        )
+
+        verifyProductExists(product.isPresent, productId).onFailure {
+            return@getExistingProduct Err(it)
+        }
+
+        return Ok(product.get())
     }
 }
